+++ wdae/wdae/users_api/management/commands/dataset_mixin.py
@@ -15,17 +15,28 @@ class DatasetBaseMixin(DatasetHelpers):
     def get_dataset(self, dataset_id):
         dataset = None
         try:
+            genotype_data = self.gpf_instance.get_genotype_data(dataset_id)
+            if genotype_data is None:
+                logger.info(f"genotype data {dataset_id} not found")
+                return None
+
             dataset = Dataset.objects.get(dataset_id=dataset_id)
         except Exception as ex:
             logger.debug(f"dataset {dataset_id} not found: {ex}")
         return dataset
 
-    def rename_wdae_dataset_and_groups(self, dataset_id, new_id):
+    def rename_wdae_dataset_and_groups(
+            self, dataset_id, new_id, dry_run=None):
+
         dataset = self.get_dataset(dataset_id)
         assert dataset is not None, f"can't find WDAE dataset {dataset_id}"
 
-        dataset.dataset_id = new_id
-        dataset.save()
+        logger.info(
+            f"going to rename wdae dataset group name from {dataset_id} "
+            f"to {new_id}")
+        if not dry_run:
+            dataset.dataset_id = new_id
+            dataset.save()
 
         new_groups = Group.objects.filter(Q(name=new_id))
         if len(new_groups) > 0:
@@ -38,16 +49,25 @@ class DatasetBaseMixin(DatasetHelpers):
         assert len(groups) == 1
         group = groups[0]
 
-        group.name = new_id
-        group.save()
+        logger.info(
+            f"going to rename wdae dataset group name from {dataset_id} "
+            f"to {new_id}")
+        if not dry_run:
+            group.name = new_id
+            group.save()
 
-    def remove_wdae_dataset_and_groups(self, dataset_id):
+    def remove_wdae_dataset_and_groups(self, dataset_id, dry_run=None):
         dataset = self.get_dataset(dataset_id)
         assert dataset is not None, f"can't find WDAE dataset {dataset_id}"
-        dataset.delete()
+        logger.info(f"going ro remove wdae dataset {dataset_id}")
+        if not dry_run:
+            dataset.delete()
 
         groups = Group.objects.filter(
             Q(groupobjectpermission__permission__codename="view"),
             Q(name=dataset_id))
         assert len(groups) == 1
-        groups.delete()
+
+        logger.info(f"going ro remove wdae dataset group {dataset_id}")
+        if not dry_run:
+            groups.delete()
